+++ relm4-macros/src/widget_macro/token_streams.rs
@@ -38,16 +38,17 @@ impl Widget {
         self.return_stream(&mut streams.return_fields);
 
         for prop in &self.properties.properties {
-            prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
-
             if let PropertyType::Widget(widget) = &prop.ty {
                 widget.generate_widget_tokens_recursively(streams, vis, model_type, relm4_path);
+                prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
+
                 if let Some(returned_widget) = &widget.returned_widget {
                     returned_widget
                         .generate_widget_tokens_recursively(streams, vis, model_type, relm4_path);
                 }
             } else {
                 prop.property_init_stream(&mut streams.assign_properties, &self.name, relm4_path);
+                prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
 
                 prop.view_stream(&mut streams.view, &self.name, relm4_path, false);
                 prop.track_stream(&mut streams.track, &self.name, model_type, false);
@@ -73,12 +74,13 @@ impl ReturnedWidget {
         self.return_stream(&mut streams.return_fields);
 
         for prop in &self.properties.properties {
-            prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
-
             if let PropertyType::Widget(widget) = &prop.ty {
                 widget.generate_widget_tokens_recursively(streams, vis, model_type, relm4_path);
+                prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
             } else {
                 prop.property_init_stream(&mut streams.assign_properties, &self.name, relm4_path);
+                prop.connect_widgets_stream(&mut streams.assign_properties, &self.name);
+
                 prop.connect_stream(&mut streams.connect, &self.name);
 
                 prop.view_stream(&mut streams.view, &self.name, relm4_path, false);